+++ R/bc.R
@@ -1,7 +1,7 @@
 
 #' Blank Concatenate. Parse argument as a vector of values allowing "blank separators".
 #'
-#' Accepts simple strings that do not include escapes. Separates data on whitespace, separating symbols.
+#' Separates string data on whitespace and separating symbols into an array.
 #' Words are assumed to start with '.', '_', or \code{getOption('wrapr.bc.alphabet')} (defaults to \code{[A-Za-z]} equivalent).
 #' Separating symbols are whitespace and \code{getOption('wrapr.bc.seps')} (defaults to \code{,|}).
 #'
@@ -9,6 +9,8 @@
 #' Suggested by Emil Erik Pula Bellamy Begtrup-Bright \url{https://github.com/WinVector/wrapr/issues/12}.
 #'
 #' @param s string to parse
+#' @param ... force later arguments to be set by name
+#' @param strict logical, if TRUE throw exception on confusing input
 #' @return vector of values
 #'
 #' @examples
@@ -20,7 +22,10 @@
 #'
 #' @export
 #'
-bc <- function(s) {
+bc <- function(
+  s,
+  ...,
+  strict = TRUE) {
   # check arguments
   if(!is.character(s)) {
     stop("wrapr::bc s must be character")
@@ -28,10 +33,12 @@ bc <- function(s) {
   if(length(s)!=1) {
     stop("wrapr::bc s must be length 1")
   }
+  wrapr::stop_if_dot_args(substitute(list(...)), "wrapr::bc")
 
   # tear up string
   sep_symbols <- getOption('wrapr.bc.seps')
-  single_quote_str <- "('[^']*')"
+  # single_quote_str <- "('[^']*')"  # no escapes
+  single_quote_str <- "('((\\\\.)|[^'])*')"  # with escapes
   double_quote_str <- gsub("'", '"', single_quote_str, fixed = TRUE)
   number_l <- '([+-]?(\\d)+[.]?(\\d)*([eE][+-]?(\\d)+)?)'
   number_r <- '([+-]?(\\d)*[.]?(\\d)+([eE][+-]?(\\d)+)?)'
@@ -50,23 +57,49 @@ bc <- function(s) {
   toks <- strsplit_capture(s,
                            split = pattern)
 
+  # special case length 0
+  if(length(toks)<=0) {
+    return(c())
+  }
+  # special case empty-return
+  if((length(toks)==1) && (nchar(toks[[1]])<=0)) {
+    return(c())
+  }
+
+  if(strict) {
+    # insist all regions were recognized
+    matched <- vapply(
+      seq_len(length(toks)),
+      function(i) { attr(toks[[i]], 'is_sep', exact = TRUE) },
+      logical(1))
+    if(!all(matched)) {
+      min_index = which(!matched)[[1]]
+      stop(paste0(
+        "non-matched token: ",
+        toks[min_index]))
+    }
+  }
+
   # limit down to non-separator regions
   is_waste <- paste0('^(\\s|[', sep_symbols, '])*$')
   indices <- grep(is_waste, toks, invert = TRUE)
-  # insist on non-consecutive value carrying indices
-  if(length(intersect(indices, indices-1))) {
-    min_index = min(intersect(indices, indices-1))
-    stop(paste0(
-      "missing explicit separator, not safe to return value: ",
-      toks[min_index],
-      toks[min_index + 1]))
-  }
-  got <- as.character(toks[indices])  # also dump attributes
 
   # special case length 0
-  if(length(got)<=0) {
+  if(length(indices)<=0) {
     return(c())
   }
+
+  if(strict) {
+    # insist on non-consecutive value carrying indices
+    if(length(intersect(indices, indices-1))) {
+      min_index = min(intersect(indices, indices-1))
+      stop(paste0(
+        "missing explicit separator, not safe to return value: ",
+        toks[min_index],
+        toks[min_index + 1]))
+    }
+  }
+  got <- as.character(toks[indices])  # also dump attributes
 
   # see if we can convert to logical
   logical <- NULL