+++ NEWS.md
@@ -1,7 +1,8 @@
 
-# wrapr 2.0.6 2021-01-30
+# wrapr 2.0.7 2021-01-30
 
  * Add bc() blank-separated parsing, suggested by Emil Erik Pula Bellamy Begtrup-Bright \url{https://github.com/WinVector/wrapr/issues/12}.
+ * Fix strsplit_capture empty-return type.
 
 # wrapr 2.0.6 2020-12-06